# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NSE Finance API",
    "description": "Python FastAPI replacement for NestJS 9 backend. Keeps MySQL finance_app (localhost:3306) and frontend compatibility GET /stocks?duration=... and GET /stock-exchange/Nse. New domain: portfolios / positions / trades / price_snapshots.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/quant/predict": {
      "post": {
        "tags": [
          "quant"
        ],
        "summary": "Qlib prediction",
        "description": "POST /quant/predict {symbol, horizon}\nReturns score in [-1,1], expected_return, confidence in [0,1].\nHeavy inference is gated behind liquidity check (mcap/volume).",
        "operationId": "predict_quant_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quant/backtest": {
      "post": {
        "tags": [
          "quant"
        ],
        "summary": "Qlib backtest",
        "description": "POST /quant/backtest {symbol, start, end}\nReturns cumulative_return, sharpe, max_drawdown, trades, etc.",
        "operationId": "backtest_quant_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Backtest Quant Backtest Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quant/screen/{symbol}": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Liquidity screen (gate for inference)",
        "description": "GET /quant/screen/{symbol} -> liquidity check.\nOnly run heavy inference if this passes (mcap >=500Cr, vol >=1L).",
        "operationId": "screen_quant_screen__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Screen Quant Screen  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quant/warnings/{symbol}": {
      "get": {
        "tags": [
          "quant"
        ],
        "summary": "Screener warnings: mcap, volume, impact",
        "description": "GET /quant/warnings/{symbol}?qty=1000\nChecks:\n  - mcap < 500Cr\n  - avgVol 20D < 1L (100k)\n  - impactCost > 2%\nReturns flags, warnings, illiquid flag.",
        "operationId": "warnings_quant_warnings__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "qty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Quantity to estimate impact for",
              "default": 1000,
              "title": "Qty"
            },
            "description": "Quantity to estimate impact for"
          },
          {
            "name": "ideal_price",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Ideal price for impact calc",
              "title": "Ideal Price"
            },
            "description": "Ideal price for impact calc"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Warnings Quant Warnings  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quant/execution/simulate": {
      "post": {
        "tags": [
          "quant"
        ],
        "summary": "Realistic execution simulation",
        "description": "POST /quant/execution/simulate {symbol, qty, ideal_price}\n\nSimulates what you could actually buy/sell:\n  feasible_qty = min(qty, avgVolume*0.1)\n  slippage = 0.1% + 0.5%*(qty/avgVolume)\n  market_impact via sqrt model\nReturns ideal vs realistic prices, feasible_qty, warning, illiquid_flag.",
        "operationId": "execution_simulate_quant_execution_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Execution Simulate Quant Execution Simulate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BacktestRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "examples": [
              "RELIANCE"
            ]
          },
          "start": {
            "type": "string",
            "title": "Start",
            "description": "YYYY-MM-DD",
            "examples": [
              "2024-01-01"
            ]
          },
          "end": {
            "type": "string",
            "title": "End",
            "description": "YYYY-MM-DD",
            "examples": [
              "2024-06-01"
            ]
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "start",
          "end"
        ],
        "title": "BacktestRequest"
      },
      "ExecutionRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "examples": [
              "RELIANCE"
            ]
          },
          "qty": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Qty",
            "description": "Requested quantity",
            "default": 1000
          },
          "ideal_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ideal Price",
            "description": "Ideal price; if null uses last close"
          },
          "side": {
            "type": "string",
            "title": "Side",
            "description": "buy or sell",
            "default": "buy"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "ExecutionRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol",
            "description": "NSE symbol, e.g. RELIANCE or RELIANCE.NS",
            "examples": [
              "RELIANCE"
            ]
          },
          "horizon": {
            "type": "string",
            "title": "Horizon",
            "description": "Prediction horizon, e.g. 5d, 10d, 1d",
            "default": "5d",
            "examples": [
              "5d"
            ]
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "PredictRequest"
      },
      "PredictResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "horizon": {
            "type": "string",
            "title": "Horizon"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "expected_return": {
            "type": "number",
            "title": "Expected Return"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "gated": {
            "type": "boolean",
            "title": "Gated"
          },
          "model_used": {
            "type": "string",
            "title": "Model Used"
          },
          "gate_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Reason"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "warnings": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "horizon",
          "horizon_days",
          "score",
          "expected_return",
          "confidence",
          "gated",
          "model_used"
        ],
        "title": "PredictResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}